Task finished - some other changes
Changed font-size of html element in file typography.scss. From 10px to 62.5%. If it's just set to 10px, users can't change the font-size, even if they have changed it in the brower's settings. It's a bad practice, becouse if someone has weak eyesight, they can't read the default page's font-size

Changed hover effect of button: btn--primaryOut

Addded:
box-sizing: border-box;
margin: 0;
padding: 0;
to the *, html in file typography.scss

Added type="button" to the button in Button.vue

Added a new component (PageHeader), a part of the landing page. Styled, but not RWD

Added the PageHeader component to the App.vue